# Conflicts:
#	src/main/resources/application.properties

+++ src/main/java/com/acme/keeplo/platform/subscription/infrastructure/persistence/jpa/PaymentCardRepository.java
@@ -5,9 +5,25 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 import java.util.Optional;
-
+/**
+ * Repository interface for accessing payment card data.
+ *
+ * Provides methods for querying PaymentCard entities from the database.
+ */
 public interface PaymentCardRepository extends JpaRepository<PaymentCard, Long> {
+
+    /**
+     * Retrieves all payment cards from the database.
+     *
+     * @return a list of all PaymentCard entities
+     */
     List<PaymentCard> findAll();
 
-    Optional<PaymentCard> findById(Integer id);
+    /**
+     * Retrieves a payment card by its ID.
+     *
+     * @param id the ID of the payment card
+     * @return an Optional containing the PaymentCard if found, or empty if not found
+     */
+    Optional<PaymentCard> findById(Long id);
 }
